solving is now possible
you can save polyominoes from the Constructor to the Palette, design a Field, and the solving algorithm will actually produce solutions in the form of Polyfits. the algorithm is very inefficient but there are plenty of ways to improve upon this. implementing the bank, creating generators, finishing saving and loading, and displaying solutions and polyomino icons are all due down the line, but the basic functionality of PolyominoSolver finally exists.